+++ azure/Kqlmagic/kql_magic.py
@@ -70,45 +70,52 @@ class Kqlmagic(Magics, Configurable):
         0, 
         config=True, 
         allow_none=True, 
-        help="Automatically limit the size of the returned result sets. Abbreviation: al"
+        help="""Automatically limit the size of the returned result sets.\n
+        Abbreviation: al"""
     )
 
     prettytable_style = Enum(
         ["DEFAULT", "MSWORD_FRIENDLY", "PLAIN_COLUMNS", "RANDOM"],
         "DEFAULT",
         config=True,
-        help="Set the table printing style to any of prettytable's defined styles. Abbreviation: ptst",
+        help="""Set the table printing style to any of prettytable's defined styles.\n
+        Abbreviation: ptst"""
     )
 
     short_errors = Bool(
         True, 
         config=True, 
-        help="Don't display the full traceback on KQL Programming Error. Abbreviation: se"
+        help="""Don't display the full traceback on KQL Programming Error.\n
+        Abbreviation: se"""
     )
 
     display_limit = Int(
         None,
         config=True,
         allow_none=True,
-        help="Automatically limit the number of rows displayed (full result set is still stored). Abbreviation: dl",
+        help="""Automatically limit the number of rows displayed (full result set is still stored).\n
+        Abbreviation: dl""",
     )
 
     auto_dataframe = Bool(
         False, 
         config=True, 
-        help="Return Pandas dataframe instead of regular result sets. Abbreviation: ad"
+        help="""Return Pandas dataframe instead of regular result sets.\n
+        Abbreviation: ad"""
     )
 
     columns_to_local_vars = Bool(
         False, 
         config=True, 
-        help="Return data into local variables from column names. Abbreviation: c2lv"
+        help="""Return data into local variables from column names.\n
+        Abbreviation: c2lv"""
     )
 
     feedback = Bool(
         True, 
         config=True, 
-        help="Show number of records returned, and assigned variables. Abbreviation: f"
+        help="""Show number of records returned, and assigned variables.\n
+        Abbreviation: f"""
     )
 
     show_conn_info = Enum(
@@ -116,64 +123,70 @@ class Kqlmagic(Magics, Configurable):
         "current",
         config=True,
         allow_none=True,
-        help="Show connection info, either current, the whole list, or None. Abbreviation: sci",
+        help="""Show connection info, either current, the whole list, or None.\n
+        Abbreviation: sci"""
     )
 
     dsn_filename = Unicode(
         "odbc.ini",
         config=True,
-        help="Path to DSN file. "
-        "When the first argument is of the form [section], "
-        "a kql connection string is formed from the "
-        "matching section in the DSN file. Abbreviation: dl",
+        allow_none=True,
+        help="""Sets path to DSN file.\n
+        When the first argument of the connection string is of the form [section], a kql connection string is formed from the matching section in the DSN file.\n
+        Abbreviation: dl"""
     )
 
     cloud = Enum(
         [Cloud.PUBLIC, Cloud.MOONCAKE, Cloud.FAIRFAX, Cloud.BLACKFOREST, Cloud.USNAT, Cloud.USSEC, Cloud.TEST],
         Cloud.PUBLIC,
         config=True,
-        help="Default cloud "
-        "the kql connection will use the cloud as specified "
+        help="""Default cloud\n"
+        "the kql connection will use the cloud as specified"""
     )
 
     enable_sso = Bool(
         False, 
         config = True, 
-        help=f"Enables or disables SSO. if enabled, SSO will only work if the environment parameter {Constants.MAGIC_CLASS_NAME.upper()}_SSO_ENCRYPTION_KEYS is set properly"
+        help=f"""Enables or disables SSO.\n
+        If enabled, SSO will only work if the environment parameter {Constants.MAGIC_CLASS_NAME.upper()}_SSO_ENCRYPTION_KEYS is set properly."""
     )
 
     sso_db_gc_interval = Int(
         168, 
         config=True,
-        help= "Garbage Collection interval for not changed SSO cache entries. Default is one week."
+        help= """Garbage Collection interval for not changed SSO cache entries. Default is one week."""
     )
 
     device_code_login_notification = Enum(
         ["frontend", "browser", "terminal", "email"],
         "frontend", 
         config = True, 
-        help = "Set device_code login notification method, default: frontend. Abbreviation: dcln"
+        help = """Sets device_code login notification method.\n
+        Abbreviation: dcln"""
     )
 
     device_code_notification_email = Unicode(
         "", 
         config=True, 
-        help=f"""Email details. Should be set by {Constants.MAGIC_CLASS_NAME.upper()}_DEVICE_CODE_NOTIFICATION_EMAIL. Abbreviation: dcne
-        the email details string format is: SMTPEndPoint='endpoint';SMTPPort='port';sendFrom='email';sendFromPassword='password';sendTo='email';context='text'"""
+        help=f"""Email details. Should be set by {Constants.MAGIC_CLASS_NAME.upper()}_DEVICE_CODE_NOTIFICATION_EMAIL.\n
+        the email details string format is: SMTPEndPoint='endpoint';SMTPPort='port';sendFrom='email';sendFromPassword='password';sendTo='email';context='text'\n
+        Abbreviation: dcne"""
     )
 
     timeout = Int(
         None, 
         config=True, 
         allow_none=True, 
-        help="Specifies the maximum time in seconds, to wait for a query response. None, means default http wait time. Abbreviation: to, wait"
+        help="""Specifies the maximum time in seconds, to wait for a query response. None, means default http wait time.\n
+        Abbreviation: to, wait"""
     )
 
     plot_package = Enum(
         ["None", "plotly", "plotly_orca"], 
         "plotly", 
         config=True, 
-        help="Set the plot package (plotlt_orca requires plotly orca to be installed on the server). Abbreviation: pp"
+        help="""Set the plot package (plotlt_orca requires plotly orca to be installed on the server).\n 
+        Abbreviation: pp"""
     )
 
     table_package = Enum(
@@ -186,43 +199,50 @@ class Kqlmagic(Magics, Configurable):
     last_raw_result_var = Unicode(
         "_kql_raw_result_", 
         config=True, 
-        help="Set the name of the variable that will contain last raw result. Abbreviation: var"
+        help="""Set the name of the variable that will contain last raw result.\n
+        Abbreviation: var"""
     )
 
     enable_suppress_result = Bool(
         True, 
         config=True, 
-        help="Suppress result when magic ends with a semicolon ;. Abbreviation: esr"
+        help="""Suppress result when magic ends with a semicolon ;.\n 
+        Abbreviation: esr"""
     )
 
     show_query_time = Bool(
         True, 
         config=True, 
-        help="Print query execution elapsed time. Abbreviation: sqt"
+        help="""Print query execution elapsed time.\n 
+        Abbreviation: sqt"""
     )
 
     show_query = Bool(
         False, 
         config=True, 
-        help="Print parametrized query. Abbreviation: sq"
+        help="""Print parametrized query.\n
+        Abbreviation: sq"""
     )
 
     show_query_link = Bool(
         False, 
         config=True, 
-        help="Show query deep link as a button, to run query in the deafult tool. Abbreviation: sql"
+        help="""Show query deep link as a button, to run query in the deafult tool.\n
+        Abbreviation: sql"""
     )
 
     query_link_destination = Enum(
         ["Kusto.Explorer", "Kusto.WebExplorer"], 
         "Kusto.WebExplorer", 
-        config=True, help="Set the deep link destination. Abbreviation: qld"
+        config=True, help="""Set the deep link destination.\n
+        Abbreviation: qld"""
     )
 
     plotly_fs_includejs = Bool(
         False,
         config=True,
-        help="Include plotly javascript code in popup window. If set to False (default), it download the script from https://cdn.plot.ly/plotly-latest.min.js. Abbreviation: pfi",
+        help="""Include plotly javascript code in popup window. If set to False (default), it download the script from https://cdn.plot.ly/plotly-latest.min.js.\n
+        Abbreviation: pfi"""
     )
 
     validate_connection_string = Bool(
@@ -234,44 +254,49 @@ class Kqlmagic(Magics, Configurable):
     auto_popup_schema = Bool(
         True, 
         config=True, 
-        help="Popup schema when connecting to a new database. Abbreviation: aps"
+        help="""Popup schema when connecting to a new database.\n
+        Abbreviation: aps"""
     )
 
     json_display = Enum(
         ["raw", "native", "formatted"], 
         "formatted", 
         config=True, 
-        help="Set json/dict display format. Abbreviation: jd"
+        help="""Set json/dict display format.\n
+        Abbreviation: jd"""
     )
 
     palette_name = Unicode(
         Palettes.DEFAULT_NAME, 
         config=True, 
-        help="Set pallete by name to be used for charts. Abbreviation: pn"
+        help="""Set pallete by name to be used for charts.\n
+        Abbreviation: pn"""
     )
 
     palette_colors = Int(
         Palettes.DEFAULT_N_COLORS, 
         config=True, 
-        help="Set pallete number of colors to be used for charts. Abbreviation: pc"
+        help="""Set pallete number of colors to be used for charts.\n
+        Abbreviation: pc"""
     
     )
     palette_desaturation = Float(
         Palettes.DEFAULT_DESATURATION, 
         config=True, 
-        help="Set pallete desaturation to be used for charts. Abbreviation: pd"
+        help="""Set pallete desaturation to be used for charts.\n
+        Abbreviation: pd"""
     )
 
     temp_folder_name = Unicode(
         f"{Constants.MAGIC_CLASS_NAME}_temp_files", 
         config=True, 
-        help="Set the folder name for temporary files"
+        help="""Set the folder name for temporary files"""
     )
 
     export_folder_name = Unicode(
         f"{Constants.MAGIC_CLASS_NAME}_exported_files", 
         config=True, 
-        help="Set the folder name  for exported files"
+        help="""Set the folder name  for exported files"""
     )
 
     cache_folder_name = Unicode(
@@ -285,57 +310,84 @@ class Kqlmagic(Magics, Configurable):
         ["auto", "jupyterlab", "jupyternotebook", "ipython", "visualstudiocode"], 
         "auto", 
         config=True, 
-        help="Set notebook application used."
+        help="""Set notebook application used."""
     ) #TODO: add "papermill"
 
     test_notebook_app = Enum(
         ["none", "jupyterlab", "jupyternotebook", "ipython", "visualstudiocode"], 
         "none", 
         config=True, 
-        help="Set testing application mode, results should return for the specified notebook application."
+        help="""Set testing application mode, results should return for the specified notebook application."""
     ) #TODO: add "papermill"
 
     add_kql_ref_to_help = Bool(
         True, 
         config=True, 
-        help=f"On {Constants.MAGIC_CLASS_NAME} load, auto add kql reference to Help menu."
+        help=f"""On {Constants.MAGIC_CLASS_NAME} load, auto add kql reference to Help menu."""
     )
 
     add_schema_to_help = Bool(
         True, 
         config=True, 
-        help="On connection to database@cluster add  schema to Help menu."
+        help="""On connection to database@cluster add  schema to Help menu."""
     )
 
     cache = Unicode(
         None, 
         config=True, 
         allow_none=True, 
-        help="Cache query results to the specified folder."
+        help="""Cache query results to the specified folder."""
     )
 
     use_cache = Unicode(
         None, 
         config=True, 
         allow_none=True, 
-        help="Use cached query results from the specified folder, instead of executing the query."
+        help="""Use cached query results from the specified folder, instead of executing the query."""
     )
 
     check_magic_version = Bool(
         True, 
         config=True, 
-        help=f"On {Constants.MAGIC_CLASS_NAME} load, check whether new version of {Constants.MAGIC_CLASS_NAME} exist"
+        help=f"""On {Constants.MAGIC_CLASS_NAME} load, check whether new version of {Constants.MAGIC_CLASS_NAME} exist"""
     )
     show_what_new = Bool(
         True, 
         config=True, 
-        help=f"On {Constants.MAGIC_CLASS_NAME} load, get history file of {Constants.MAGIC_CLASS_NAME} and show what new button to open it"
+        help=f"""On {Constants.MAGIC_CLASS_NAME} load, get history file of {Constants.MAGIC_CLASS_NAME} and show what new button to open it"""
     )
 
     show_init_banner = Bool(
         True, 
         config=True, 
-        help=f"On {Constants.MAGIC_CLASS_NAME} load, show init banner"
+        help=f"""On {Constants.MAGIC_CLASS_NAME} load, show init banner"""
+    )
+
+    request_id_tag = Unicode(
+        None, 
+        config=True, 
+        allow_none=True, 
+        help=f"""Tags request 'x-ms-client-request-id' header.\n
+        Header pattern: {Constants.MAGIC_CLASS_NAME}.execute;{{tag}};{{guid}}\n
+        Abbreviation: idtag"""
+    )
+
+    request_app_tag = Unicode(
+        None, 
+        config=True, 
+        allow_none=True, 
+        help=f"""Tags request 'x-ms-app' header.\n
+        Header pattern: {Constants.MAGIC_CLASS_NAME};{{tag}}\n
+        Abbreviation: apptag"""
+    )
+
+    request_user_tag = Unicode(
+        None, 
+        config=True, 
+        allow_none=True, 
+        help=f"""Tags request 'x-ms-user' header.\n
+        Header pattern: {{tag}}\n
+        Abbreviation: usertag"""
     )
 
     logger().debug("Kqlmagic:: - define class code")